.NET 7 & ImageSharp 3 Update
