+++ .github/workflows/dotnet.yml
@@ -14,13 +14,17 @@ jobs:
         os: [ ubuntu-latest, windows-latest ]
     steps:
       - uses: actions/checkout@v2
-      - name: Setup .NET
+      - name: Setup .NET 6.0
         uses: actions/setup-dotnet@v1
         with:
           dotnet-version: 6.0.x
+      - name: Setup .NET 7.0
+        uses: actions/setup-dotnet@v1
+        with:
+          dotnet-version: 7.0.x
       - name: Restore dependencies
         run: dotnet restore
       - name: Build
         run: dotnet build --no-restore
       - name: Test
-        run: dotnet test --no-build --verbosity normal DMISharpTests/DMISharpTests.csproj
+        run: dotnet test --no-build --verbosity normal DMISharp.Tests/DMISharp.Tests.csproj
