+++ DMISharp.Benchmark/Data/Input/352x352.dmi
