+++ DMISharp.Benchmark/Data/Input/animal.dmi
