+++ DMISharp.Benchmark/Data/Input/barsigns.dmi
