+++ DMISharp.Benchmark/Data/Input/large.dmi
