+++ DMISharp.Benchmark/Data/Input/small.dmi
