+++ DMISharp.Benchmark/Program.cs
@@ -1,0 +1,14 @@
+﻿
+using BenchmarkDotNet.Running;
+
+namespace DMISharp.Benchmark;
+
+public static class Program
+{
+    public static void Main(string[] args)
+    {
+        _ = BenchmarkSwitcher
+            .FromAssembly(typeof(Program).Assembly)
+            .Run(args);
+    }
+}
