+++ DMISharp.Tests/Data/Input/352x352.dmi
