+++ DMISharp.Tests/Data/Input/SourceImages/sord.png
