+++ DMISharp.Tests/Data/Input/SourceImages/sordvert.png
