+++ DMISharp.Tests/Data/Input/SourceImages/steve32.png
