+++ DMISharp.Tests/Data/Input/air_meter.dmi
