+++ DMISharp.Tests/Data/Input/animal.dmi
