+++ DMISharp.Tests/Data/Input/atmos_testing.dmi
