+++ DMISharp.Tests/Data/Input/barsigns.dmi
