+++ DMISharp.Tests/Data/Input/broadMobs.dmi
