+++ DMISharp.Tests/Data/Input/light_64.dmi
