+++ DMISharp.Tests/Data/Input/lighting.dmi
