+++ DMISharp.Tests/Data/Input/misc.dmi
