+++ DMISharp.Tests/Data/Input/spacedragon.dmi
