+++ DMISharp.Tests/Data/Input/turf_analysis_goon.dmi
