+++ DMISharp/DMISharp.csproj
@@ -1,45 +1,43 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFrameworks>net6.0;netstandard2.0;net461;net472</TargetFrameworks>
-    <Authors>Bobbahbrown</Authors>
-    <Company>MelonMesa</Company>
-    <Description>Library for handling BYOND DMI files.</Description>
-    <Copyright>Copyright (c) 2022 MelonMesa</Copyright>
-    <PackageProjectUrl>https://melonmesa.com</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/bobbahbrown/DMISharp</RepositoryUrl>
-    <RepositoryType>Git</RepositoryType>
-    <AssemblyVersion>1.7.0.0</AssemblyVersion>
-    <Version>1.7.0</Version>
-    <PackageIcon>dmisharp.png</PackageIcon>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageLicenseExpression>GPL-3.0-only</PackageLicenseExpression>
-    <PackageTags>DMI BYOND SpaceStation13</PackageTags>
-    <PackageReleaseNotes>
-      - Added compression optimizations and some other size-decreasing optimizations to DMI file saving
-      - Updated to .NET 6
-      - Updated ImageSharp to 2.0
-    </PackageReleaseNotes>
-    <FileVersion>1.7.0.0</FileVersion>
-    <LangVersion>8.0</LangVersion>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFrameworks>net6.0;net7.0</TargetFrameworks>
+        <Authors>Bobbahbrown</Authors>
+        <Company>MelonMesa</Company>
+        <Description>Library for handling BYOND DMI files.</Description>
+        <Copyright>Copyright (c) 2023 MelonMesa</Copyright>
+        <PackageProjectUrl>https://github.com/bobbahbrown/DMISharp</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/bobbahbrown/DMISharp</RepositoryUrl>
+        <RepositoryType>Git</RepositoryType>
+        <AssemblyVersion>1.8.0.0</AssemblyVersion>
+        <Version>1.8.0</Version>
+        <FileVersion>1.8.0.0</FileVersion>
+        <PackageIcon>dmisharp.png</PackageIcon>
+        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <PackageLicenseExpression>GPL-3.0-only</PackageLicenseExpression>
+        <PackageTags>DMI BYOND SpaceStation13</PackageTags>
+        <PackageReleaseNotes>
+            - Updated to .NET 7
+            - Updated ImageSharp to 3
+        </PackageReleaseNotes>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="MetadataExtractor" Version="2.7.1" />
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.2">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Toolkit.HighPerformance" Version="7.1.2" />
-    <PackageReference Include="SixLabors.ImageSharp" Version="2.0.0" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="MetadataExtractor" Version="2.8.0" />
+        <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.3">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="Microsoft.Toolkit.HighPerformance" Version="7.1.2" />
+        <PackageReference Include="SixLabors.ImageSharp" Version="3.0.1" />
+    </ItemGroup>
 
-  <ItemGroup Condition="$(TargetFramework.StartsWith('net4')) OR $(TargetFramework.StartsWith('netstandard'))">
-    <PackageReference Include="IndexRange" Version="1.0.1" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <None Include="NugetIcon\dmisharp.png" Pack="true" PackagePath="" />
-  </ItemGroup>
+    <ItemGroup Condition="$(TargetFramework.StartsWith('net4')) OR $(TargetFramework.StartsWith('netstandard'))">
+        <PackageReference Include="IndexRange" Version="1.0.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Include="NugetIcon\dmisharp.png" Pack="true" PackagePath="" />
+    </ItemGroup>
 
 </Project>
