--- DMISharp/GlobalSuppressions.cs
@@ -1,3 +1,0 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage("Globalization", "CA1303:Do not pass literals as localized parameters", Justification = "<Pending>", Scope = "member", Target = "~M:DMISharp.DMIFile.Save(System.IO.Stream)")]
