--- DMISharpBenchmark/Data/Output/EmptyFile.txt
@@ -1,1 +1,0 @@
-﻿I'm empty for a reason!
