--- DMISharpBenchmark/Program.cs
@@ -1,14 +1,0 @@
-﻿using BenchmarkDotNet.Running;
-
-namespace DMISharpBenchmark
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            _ = BenchmarkSwitcher
-                .FromAssembly(typeof(Program).Assembly)
-                .Run(args);
-        }
-    }
-}
